Add docstrings

+++ preprocessing.py
@@ -1,6 +1,11 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
+"""
+Preprocessing procedures that transform data from cycles.csv and tracking.csv
+into the sequence of sequences required by RNNs.
+"""
+
 import os
 from os.path import join as pj
 
@@ -61,28 +66,71 @@ training_columns = list_of_symptoms + ['day_in_cycle', 'absolute_day', 'period']
 
 # ============  Feature engineer for the cycle ============
 def expand_cycle(cycle):
+    """Expand information about a given cycle for a given user.
+
+    Parameters
+    ----------
+    cycle : pd.Series
+        A single row from cycles.csv
+
+    Returns
+    -------
+    expanded_cycle : pd.DataFrame
+        DataFrame with expanded information about the given cycle
+        - It has a pd.MultiIndex with the user_id and a date range spanning through
+          the days of the given cycle
+        - It has as columns the `cycle_id`, `day_in_cycle` and `period`. The last one
+          being a boolean indicating days of period.
+    """
+
+    # Get date range for the cycle
     dates = pd.date_range(start=cycle.cycle_start, periods=cycle.cycle_length).tolist()
+    # Create a boolean indicator of period days
     period = np.zeros(cycle.cycle_length, dtype=np.int8)
     period[:cycle.period_length] = 1
+    # Enumerate days in cycle
     day_in_cycle = np.arange(1, cycle.cycle_length + 1, dtype=np.int8)
 
+    # Build the index out of user_id and dates
     index = pd.MultiIndex.from_tuples(
         tuples=list(zip([cycle.user_id] * int(cycle.cycle_length), dates)),
         names=["user_id", "date"]
     )
 
-    return pd.DataFrame(
+    # Build up the DataFrame
+    expanded_cycle = pd.DataFrame(
         data=list(zip([cycle.cycle_id] * int(cycle.cycle_length), day_in_cycle, period)),
         index=index,
         columns=['cycle_id', 'day_in_cycle', 'period']
     )
 
+    return expanded_cycle
+
 
 def expand_cycles(cycles):
+    """Expand all cycles for all users.
+
+    Simple iterator over exapnd_cycles. The output of this function has two purposes:
+    - It brings period days as a feature to the final `features` DataFrame.
+    - When merging it with tracking information, it adds up empty inactive days to the
+      final `features` DataFrame.
+
+    Parameters
+    ----------
+    cycles : pd.DataFrame
+        DataFrame loaded from cycles.csv
+
+    Returns
+    -------
+    cycles_processed : pd.DataFrame
+        Complete daily details per user and cycle
+    """
+
     cycles_processed_backup = pj(staging_dir, "cycles_processed.pkl.gz")
-    try:
+    # Try to load from memory if already computed
+    if os.path.exists(cycles_processed_backup):
         cycles_processed = joblib.load(cycles_processed_backup)
-    except:
+    else:
         cycles_processed = pd.concat([expand_cycle(cycle) for _, cycle in cycles.iterrows()])
         joblib.dump(cycles_processed, cycles_processed_backup)
 
@@ -91,7 +139,22 @@ def expand_cycles(cycles):
 
 # ========  Feature engineer for tracked symptoms =========
 def process_tracking(tracking):
-    # One hot encode the symptoms
+    """One hot encode the symptoms.
+
+    Parameters
+    ----------
+    tracking : pd.DataFrame
+        DataFrame loaded from tracking.csv
+
+    Returns
+    -------
+    tracking_processed : pd.DataFrame
+        Aggregated information about symptoms logged per active day.
+        - The DataFrame has one line per user and per active day.
+        - Columns are boolean indicators of logged symptoms ordered according to
+          `list_of_symptoms`
+
+    """
     tracking_processed = pd.get_dummies(
         tracking[["user_id", "date", "symptom"]],
         columns=['symptom'], prefix='', prefix_sep=''
@@ -103,6 +166,39 @@ def process_tracking(tracking):
 
 # ===============  Merging all the features ===============
 def get_features(split=True, force=False):
+    """Extract features from data.
+
+    Transforms information on `cycles.csv` and `tracking.csv` into a pandas DataFrame such
+    that there is a line per user, per day since the day she started using the app.
+    including inactive days is important because we need to keep the notion of time. The
+    RNN also needs to learn inactivity.
+
+
+    Parameters
+    ----------
+    split : Bool
+        If True, split into train/test for the training step of the model
+
+    force : Bool
+        To speed-up development, a backup of features is created after the first call.
+        After the second call, the backup will be loaded and intermediary steps are skipped.
+        If the data/function changes or you want to recompute the features, force=True
+        will ignore the backup.
+
+    Returns
+    -------
+    features : pd.DataFrame
+        DataFrame holding the features to be used for predicting.
+        - For each user, there is a row per date since the day she starting using the app
+          until the end of her last period.
+        - The first 81 columns are booleans indicating if the user had a given symptom on
+          a given day. The ordering of the symptoms goes according to `list_of_symptoms`.
+          The last 3 columns correspond to `cycle_id`, `day_in_cycle` and `period`. `period`
+          is a boolean indicating if the user had her period that day.
+        - Inactive days also get a row on this DataFrame for which symptoms are filled with
+          zeros and `cycle_id`, `day_in_cycle` and `period` are properly backfilled from the
+          information in cycles.
+    """
 
     features_backup = pj(staging_dir, 'features.pkl.gz')
 
@@ -157,16 +253,34 @@ def get_features(split=True, force=False):
 
 # =========== Prepare data for model predictions ==========
 def prepare_data_for_prediction(features=None, maxlen=90):
-    """Clip the training dataset to maxlen days per user
+    """Clip the training dataset to maxlen days per user.
 
-    The LSTM predictor uses only the last maxlen days per user for predicting.
+    The LSTM predictor uses only the last `maxlen` days per user for predicting. This function
+    reindexes the features dataframe so that only the last `maxlen` days of activity
+    of each user are considered. If a user has logged has been active less than `maxlen` days,
+    it will backfill with 0 to comply with the specifications of the model.
+
+    Parameters
+    ----------
+    features : pd.DataFrame
+        DataFrame as provided by get_features (Refer to get_features docstring).
+        If None, get_features() is called inside the function.
+
+    maxlen : int (default=90)
+        Number of days used as features for predicting the next cycle
+
+    Returns
+    -------
+    features : pd.DataFrame
+        Clipped features with only the last `maxlen` days (back-filled with zeros if necessary)
+
     """
 
     def get_user_daterange_index(user_id, max_date, length):
         """Return a list of tuples with [user_id] x list of 'length' last days.
 
         The list of tuples will be used to reconstruct an index with only the
-        the previous 90 dates per user
+        the previous `length` dates per user starting from 'max_date' backwards.
         """
         dates = pd.date_range(end=max_date, periods=length)
         return list(zip([user_id] * len(dates), dates))
